+++ src/framebuffer/screeninfo.rs
@@ -48,7 +48,7 @@ pub struct VarScreeninfo {
 #[derive(Clone, Debug)]
 pub struct FixScreeninfo {
     pub id: [u8; 16],
-    pub smem_start: usize,
+    pub smem_start: u32,
     pub smem_len: u32,
     pub fb_type: u32,
     pub type_aux: u32,
@@ -57,7 +57,7 @@ pub struct FixScreeninfo {
     pub ypanstep: u16,
     pub ywrapstep: u16,
     pub line_length: u32,
-    pub mmio_start: usize,
+    pub mmio_start: u32,
     pub mmio_len: u32,
     pub accel: u32,
     pub capabilities: u16,
